Limit Apps menu with link to All Apps
The Apps menu displays pinned apps, and currently there is no limit to the
number of pinned apps that would display.

This change limits the number of pinned apps displayed in the menu to a
configurable limit and provides a separate link to the All Apps page in case a
user wants to view the rest of the apps.

The pinned apps widget on the home page of the dashboard will always display
all of the pinned apps.