show shared apps on the dashboard
This removes the toggle that sets the root to apps#featured when app
sharing is enabled. dashboard#index is now always the root. This commit
has half backward compatability in that users don't need to set pinned_apps
for the apps to show up on the dashboard. But they will need to set
pinned_apps_group_by to category in order to get the same look and as 1.8-
where each app is in it's own category.