refactor fa icon info into a webpack (#1377)
The _form_icon.html.erb in this change was dos encoded and I switched it to unix so that's why it looks as if it's changed more than it has. What has really changed in it is only removing the script element to a pack and adding the pack_tag.