projects to use numeric ids instead of directories (#2601)
projects to use numeric ids instead of directories

projects to use numeric ids instead of directories for a
variety of reasons. Notably to decouple the URLs from the
initial directory name (opening us up to allow for spaces).
What's more is it'll allow us to use directories _outside_
of the dataroot because the full path is in the lookup file.