save excludable options (#2929)
* save excludable options

This enhances excludable options so that all options are shown in
scripts#edit when you're choosing which to show and which to hide.
Then in scripts#show the options you've chosen to exclude are indeed
hidden.  This also adds a new field to attribute so that we can
save which options to exclude.